portal-backend/requirements.txt
Flask>=2.3.0
psycopg2-binary>=2.9.0
flask-cors
flask-sqlalchemy
tenacity
flask-restful
flasgger
pytest>=7.0.0
pytest-flask>=1.2.0